Cambiamos el formato del campo frequency por FLOAT

+++ database/migrations/2016_03_19_050846_create_radios_table.php
@@ -21,7 +21,7 @@ class CreateRadiosTable extends Migration
             $table->string('name');
             $table->integer('frequency');
             $table->string('streaming');
-            $table->integer('active');
+            $table->float('active');
             $table->timestamps();
         });
     }
